Implemented instantiation in single class
