+++ src/main/java/org/opentestsystem/ap/migration/migration/Migration1615.java
@@ -21,7 +21,7 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import java.lang.reflect.Constructor;
+
 import java.util.Set;
 
 @Component
@@ -34,16 +34,15 @@ public class Migration1615 extends AbstractDataStoreMigration {
     @Override
     protected ItemEntity migrateEntity(ItemEntity itemEntity) {
         Reflections reflections = new Reflections(this.getClass().getName());
-        Set<Class<?>> fixerClasses = reflections.getTypesAnnotatedWith(Migration1615Fixer.FixFor.class);
+        Set<Class<?>> fixerClasses = reflections.getTypesAnnotatedWith(FixFor.class);
 
         try {
             for (Class<?> clazz : fixerClasses) {
                 if (itemEntity.getItemJson().getType()
-                        .equals(clazz.getAnnotation(Migration1615Fixer.FixFor.class).type())) {
-//                    Migration1615Fixer.EbsrItemFixer fixer2 = new Migration1615Fixer.EbsrItemFixer();
-//                    Constructor constr = clazz.getConstructor();
-                    Migration1615Fixer.HtmlFixer fixer = (Migration1615Fixer.HtmlFixer) clazz.getConstructor().newInstance();
-//                    fixer.fixHtml(itemEntity.getItemJson());
+                        .equals(clazz.getAnnotation(FixFor.class).type())) {
+//                  EbsrItemFixer fixer2 = new EbsrItemFixer();
+                    HtmlFixer fixer = (HtmlFixer) clazz.getConstructor().newInstance();
+                    fixer.fixHtml(itemEntity.getItemJson());
                     String a = "";
                 }
             }
@@ -54,8 +53,40 @@ public class Migration1615 extends AbstractDataStoreMigration {
         return itemEntity;
     }
 
+    @Retention(RetentionPolicy.RUNTIME)
+    @Target(ElementType.TYPE)
+    public @interface FixFor {
+        String type();
+    }
 
+    public interface HtmlFixer {
+        Item fixHtml(Item item);
+    }
 
+    @NoArgsConstructor
+    @FixFor(type = ItemConstants.ItemType.TYPE_SA)
+    public static class SaItemFixer implements HtmlFixer {
+        @Override
+        public Item fixHtml(Item item) {
+            SaItem saItem = (SaItem) item;
+            saItem.getCore().getEn().setPrompt("sa fixed");
+            saItem.getTranslations().getEsp().setPrompt("sa solucionado");
 
+            return item;
+        }
+    }
+
+    @NoArgsConstructor
+    @FixFor(type = ItemConstants.ItemType.TYPE_EBSR)
+    public static class EbsrItemFixer implements HtmlFixer {
+        @Override
+        public Item fixHtml(Item item) {
+            EbsrItem ebsrItem = (EbsrItem) item;
+            ebsrItem.getCore().getEn().setPrompt("ebsr fixed");
+            ebsrItem.getTranslations().getEsp().setPrompt("ebsr solucionado");
+
+            return item;
+        }
+    }
 
 }
