+++ src/main/java/org/opentestsystem/ap/migration/migration/Migration1615Fixer.java
@@ -13,39 +13,5 @@ import java.lang.annotation.Target;
 
 @NoArgsConstructor
 public class Migration1615Fixer {
-    @Retention(RetentionPolicy.RUNTIME)
-    @Target(ElementType.TYPE)
-    public @interface FixFor {
-        String type();
-    }
 
-    public interface HtmlFixer {
-        Item fixHtml(Item item);
-    }
-
-    @NoArgsConstructor
-    @FixFor(type = ItemConstants.ItemType.TYPE_SA)
-    public static class SaItemFixer implements HtmlFixer {
-        @Override
-        public Item fixHtml(Item item) {
-            SaItem saItem = (SaItem) item;
-            saItem.getCore().getEn().setPrompt("sa fixed");
-            saItem.getTranslations().getEsp().setPrompt("sa solucionado");
-
-            return item;
-        }
-    }
-
-    @NoArgsConstructor
-    @FixFor(type = ItemConstants.ItemType.TYPE_EBSR)
-    public static class EbsrItemFixer implements HtmlFixer {
-        @Override
-        public Item fixHtml(Item item) {
-            EbsrItem ebsrItem = (EbsrItem) item;
-            ebsrItem.getCore().getEn().setPrompt("ebsr fixed");
-            ebsrItem.getTranslations().getEsp().setPrompt("ebsr solucionado");
-
-            return item;
-        }
-    }
 }
